Added windows stub
