+++ harden/bash_log_windows.go
@@ -1,0 +1,7 @@
+package harden
+
+
+func configureBash(filePath string, backupPath string) {
+	// temp
+
+}
